Hello world example

See merge request lts/sleipnir!36

+++ Vehicle-Python/Fuel/Fuel.h
@@ -126,19 +126,19 @@ public:
      * @param _index New value for member index
      */
     eProsima_user_DllExport void index(
-            double _index);
+            uint32_t _index);
 
     /*!
      * @brief This function returns the value of member index
      * @return Value of member index
      */
-    eProsima_user_DllExport double index() const;
+    eProsima_user_DllExport uint32_t index() const;
 
     /*!
      * @brief This function returns a reference to member index
      * @return Reference to member index
      */
-    eProsima_user_DllExport double& index();
+    eProsima_user_DllExport uint32_t& index();
 
     /*!
      * @brief This function copies the value in member message
@@ -225,7 +225,7 @@ public:
 
 private:
 
-    double m_index;
+    uint32_t m_index;
     std::string m_message;
 };
